issue 3634: sagemaker
