+++ extended/src/main/java/apoc/ExtendedApocConfig.java
@@ -30,9 +30,6 @@ public class ExtendedApocConfig extends LifecycleAdapter
     public static final String APOC_UUID_ENABLED_DB = "apoc.uuid.enabled.%s";
     public static final String APOC_UUID_FORMAT = "apoc.uuid.format";
     public static final String APOC_OPENAI_KEY = "apoc.openai.key";
-    public static final String APOC_ML_OPENAI_URL = "apoc.ml.openai.url";
-    public static final String APOC_ML_OPENAI_TYPE = "apoc.ml.openai.type";
-    public static final String APOC_ML_OPENAI_AZURE_VERSION = "apoc.ml.azure.api.version";
     public static final String APOC_AWS_KEY_ID = "apoc.aws.key.id";
     public static final String APOC_AWS_SECRET_KEY = "apoc.aws.secret.key";
     public enum UuidFormatType { hex, base64 }
